properly loading wallet
